requirements.txt
appdirs==1.4.4
attrs==20.3.0
beautifulsoup4==4.9.3
blis==0.7.4
cachetools==4.2.1
catalogue==2.0.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
cssselect==1.1.0
cymem==2.0.5
dateparser==1.0.0
distlib==0.3.1
docker-pycreds==0.4.0
falcon==2.0.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
fr-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_lg-3.0.0/fr_core_news_lg-3.0.0-py3-none-any.whl
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.0.0/fr_core_news_md-3.0.0-py3-none-any.whl
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.0.0/fr_core_news_sm-3.0.0-py3-none-any.whl
gitdb==4.0.7
GitPython==3.1.14
google-alerts==0.2.9
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
GoogleNews==1.5.7
httplib2==0.19.1
idna==2.10
iniconfig==1.1.1
jieba3k==0.35.1
Jinja2==2.11.3
joblib==1.0.1
langdetect==1.0.7
lxml==4.6.3
MarkupSafe==1.1.1
murmurhash==1.0.5
newspaper3k==0.2.8
nltk==3.5
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
pathtools==0.1.2
pathy==0.4.0
Pillow==8.2.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.5
promise==2.3
protobuf==3.15.6
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.1
scipy==1.6.2
sentencepiece @ file:///Users/alex/PycharmProjects/SentinelleNLP/WheelFiles/sentencepiece-0.1.95-cp39-cp39-macosx_10_6_x86_64.whl
sentry-sdk==1.0.0
sgmllib3k==1.0.0
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smart-open==3.0.0
smmap==4.0.0
soupsieve==2.2.1
spacy==3.0.5
spacy-alignments==0.8.3
spacy-langdetect==0.1.2
spacy-legacy==3.0.2
spacy-transformers==1.0.2
srsly==2.4.1
subprocess32==3.5.4
thinc==8.0.3
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tokenizers==0.10.2
toml==0.10.2
torch==1.8.1
tqdm==4.59.0
transformers==4.5.1
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.3
wandb==0.10.24
wasabi==0.8.2
wikipedia==1.4.0
